update FTDY to the new format
