+++ buildmaster/E605/kinematics.yaml
