+++ buildmaster/E605/rawdata/HEPData-ins302822-v1-Table_6.yaml
