+++ buildmaster/E866_FTDY_800GEV_PDXSECRATIO/kinematics.yaml
