+++ buildmaster/E866_FTDY_800GEV_PDXSECRATIO/metadata.yaml
@@ -1,0 +1,49 @@
+# Generalia
+setname: "E866_FTDY_800GEV_PDXSECRATIO"
+
+version: 1
+version_comment: "Initial implementation"
+
+# References
+arXiv:
+   url: "https://arxiv.org/pdf/hep-ex/0103030.pdf"
+iNSPIRE:
+   url: "https://inspirehep.net/literature?sort=mostrecent&size=25&page=1&q=find%20eprint%20hep-ex%2F0103030"
+hepdata:
+   url: "https://www.hepdata.net/record/ins5543165"
+   version: 1
+   tables: [4]
+
+nnpdf_metadata:
+   nnpdf31_process: "DY"
+   experiment: "E866"
+
+implemented_observables:
+  - dataset: "E866_FTDY_800GEV_PDXSECRATIO"
+    observable: {description: "Data on the ratio of pd DY cross sections over pp cross sections", label: $\sigma^d/2 \sigma^p$, units: "" }
+    ndata: 15
+
+    dataset_label: E866R $d/2*p$
+    plot_x: x2
+    figure_by: 
+      - M
+    kinematic_coverage:
+      x2: x2
+      M: M
+
+    kinematics:
+      variables:
+          y: {description: "y", label: "$x_2$", units: ""}
+          M:  {description: "Dilepton mass", units: ""}
+          sqrts:  {description: "Center of mass energy", units: ""}
+      file: kinematics.yaml
+
+    data_central: data.yaml
+    data_uncertainties:
+      - uncertainties.yaml
+
+    theory:
+      FK_tables:
+        - - DYE866R_D
+        - - DYE866R_P
+      operation: 'ratio'
