+++ buildmaster/E866_FTDY_800GEV_PXSEC/kinematics.yaml
