+++ buildmaster/E866_FTDY_800GEV_PXSEC/rawdata/HEPData-ins613362-v1-Table_5.yaml
