+++ buildmaster/E906/kinematics.yaml
