+++ buildmaster/E906/rawdata/data_paper.dat
